refactor(eve): narrow subagent task admission
